// src/scenes/tips/enums/tips.enum.ts
export enum TIPS_MAIN_BUTTON_CALLBACK {
  PROTEIN = 'TIPS_MAIN_PROTEIN',
  VITAMIN_C = 'TIPS_MAIN_VITAMIN_C',
  OMEGA = 'TIPS_MAIN_OMEGA',
  CALCIUM = 'TIPS_MAIN_CALCIUM',
  IRON = 'TIPS_MAIN_IRON',
  SELENIUM = 'TIPS_MAIN_SELENIUM',
  ZINC = 'TIPS_MAIN_ZINC',
  ANTIOXIDANT = 'TIPS_MAIN_ANTIOXIDANT',
  MUSIC_AFFIRMATIONS = 'TIPS_MAIN_MUSIC_AFFIRMATIONS',
  DYNAMIC_AFFIRMATIONS = 'TIPS_MAIN_DYNAMIC_AFFIRMATIONS',
}

export enum TIPS_DYNAMIC_AFFIRMATION_BUTTON_CALLBACK {
  LUCK = 'TIPS_DYNAMIC_AFFIRMATION_LUCK',
  LOVE = 'TIPS_DYNAMIC_AFFIRMATION_LOVE',
  PASSION = 'TIPS_DYNAMIC_AFFIRMATION_PASSION',
  HEALTH = 'TIPS_DYNAMIC_AFFIRMATION_HEALTH',
  FAMILY = 'TIPS_DYNAMIC_AFFIRMATION_FAMILY',
  MONEY = 'TIPS_DYNAMIC_AFFIRMATION_MONEY',
  FEMALE = 'TIPS_DYNAMIC_AFFIRMATION_FEMALE',
  MOTHERHOOD = 'TIPS_DYNAMIC_AFFIRMATION_MOTHERHOOD',
  BEAUTY = 'TIPS_DYNAMIC_AFFIRMATION_BEAUTY',
  MIND = 'TIPS_DYNAMIC_AFFIRMATION_MIND',
  SUCCESS = 'TIPS_DYNAMIC_AFFIRMATION_SUCCESS',
  PEACE_VICTORY = 'TIPS_DYNAMIC_AFFIRMATION_PEACE_VICTORY',
}

export enum TIPS_MUSIC_AFFIRMATION_BUTTON_CALLBACK {
  MARRIAGE = 'TIPS_MUSIC_AFFIRMATION_MARRIAGE',
  FAMILY = 'TIPS_MUSIC_AFFIRMATION_FAMILY',
  FEMALE = 'TIPS_MUSIC_AFFIRMATION_FEMALE',
}